Move global current_executor functions to current_executor files
