+++ libs/threading_base/include/hpx/threading_base/thread_helpers.hpp
@@ -401,22 +401,6 @@ namespace hpx { namespace threads {
     HPX_API_EXPORT void reset_continuation_recursion_count();
     /// \endcond
 
-    /// Returns a reference to the executor which was used to create
-    /// the given thread.
-    ///
-    /// \throws If <code>&ec != &throws</code>, never throws, but will set \a ec
-    ///         to an appropriate value when an error occurs. Otherwise, this
-    ///         function will throw an \a hpx#exception with an error code of
-    ///         \a hpx#yield_aborted if it is signaled with \a wait_aborted.
-    ///         If called outside of a HPX-thread, this function will throw
-    ///         an \a hpx#exception with an error code of \a hpx::null_thread_id.
-    ///         If this function is called while the thread-manager is not
-    ///         running, it will throw an \a hpx#exception with an error code of
-    ///         \a hpx#invalid_status.
-    ///
-    HPX_API_EXPORT threads::executors::current_executor get_executor(
-        thread_id_type const& id, error_code& ec = throws);
-
     /// Returns a pointer to the pool that was used to run the current thread
     ///
     /// \throws If <code>&ec != &throws</code>, never throws, but will set \a ec
@@ -610,22 +594,6 @@ namespace hpx { namespace this_thread {
         return suspend(std::chrono::milliseconds(ms),
             threads::invalid_thread_id, description, ec);
     }
-
-    /// Returns a reference to the executor which was used to create the current
-    /// thread.
-    ///
-    /// \throws If <code>&ec != &throws</code>, never throws, but will set \a ec
-    ///         to an appropriate value when an error occurs. Otherwise, this
-    ///         function will throw an \a hpx#exception with an error code of
-    ///         \a hpx#yield_aborted if it is signaled with \a wait_aborted.
-    ///         If called outside of a HPX-thread, this function will throw
-    ///         an \a hpx#exception with an error code of \a hpx::null_thread_id.
-    ///         If this function is called while the thread-manager is not
-    ///         running, it will throw an \a hpx#exception with an error code of
-    ///         \a hpx#invalid_status.
-    ///
-    HPX_EXPORT threads::executors::current_executor get_executor(
-        error_code& ec = throws);
 
     /// Returns a pointer to the pool that was used to run the current thread
     ///
